CHeck boost_1_57_0.tar.gz Before Remove (#891)
ensure the boost_1_57_0.tar.gz exists before remove action.

+++ build.sh
@@ -26,7 +26,9 @@ cd ${DEPS_SOURCE}
 # boost
 if [ ! -f "${FLAG_DIR}/boost_1_57_0" ] \
     || [ ! -d "${DEPS_PREFIX}/boost_1_57_0/boost" ]; then
-    rm boost_1_57_0.tar.gz 
+    if [ -e boost_1_57_0.tar.gz  ]; then
+        rm boost_1_57_0.tar.gz 
+    fi 
     wget https://raw.githubusercontent.com/lylei9/boost_1_57_0/master/boost_1_57_0.tar.gz
     tar zxf boost_1_57_0.tar.gz
     rm -rf ${DEPS_PREFIX}/boost_1_57_0
